+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "editor.rulers": [51]
+  "editor.rulers": [80]
 }
