+++ baseline/OSPS-LE.yaml
@@ -1,108 +1,115 @@
-- id: OSPS-LE-01
-  maturity_level: 2
-  category: Legal
-  criterion: |
-    The version control system MUST require all
-    code contributors to assert that they are
-    legally authorized to commit the associated
-    contributions on every commit.
-  rationale: |
-    Ensure that code contributors are aware of
-    and acknowledge their legal responsibility
-    for the contributions they make to the
-    project, reducing the risk of intellectual
-    property disputes.
-  details: |
-    Include a DCO or CLA in the project's
-    repository, requiring code contributors to
-    assert that they are legally authorized to
-    commit the associated contributions on every
-    commit. Use a status check to ensure the
-    assertion is made.
-  control_mappings: # TODO
-  security_insights_value: # TODO
+category: Legal
+description: |
+  Legal criteria focus on the policies and
+  procedures that govern the project's licensing
+  and intellectual property. These criteria help
+  ensure that the project's source code is
+  distributed under a recognized and legally
+  enforceable open source software license,
+  reducing the risk of intellectual property
+  disputes or licensing violations.
+criteria:
+  - id: OSPS-LE-01
+    maturity_level: 2
+    criterion: |
+      The version control system MUST require all
+      code contributors to assert that they are
+      legally authorized to commit the associated
+      contributions on every commit.
+    rationale: |
+      Ensure that code contributors are aware of
+      and acknowledge their legal responsibility
+      for the contributions they make to the
+      project, reducing the risk of intellectual
+      property disputes against the project.
+    details: |
+      Include a DCO or CLA in the project's
+      repository, requiring code contributors to
+      assert that they are legally authorized to
+      commit the associated contributions on every
+      commit. Use a status check to ensure the
+      assertion is made.
+    control_mappings: # TODO
+    security_insights_value: # TODO
 
-- id: OSPS-LE-02
-  maturity_level: 1
-  category: Legal
-  criterion: |
-    The license for the source code MUST
-    meet the OSI Open Source Definition or
-    the FSF Free Software Definition.
-  rationale: |
-    Ensure that the project's source code is
-    distributed under a recognized and legally
-    enforceable open source software license,
-    providing clarity on how the code
-    can be used and shared by others.
-  details: |
-    Add a LICENSE file to the project's repo
-    with a license that is
-    an approved license by the
-    Open Source Initiative (OSI), or
-    a free license as approved by the
-    Free Software Foundation (FSF).
-    Examples of such licenses include the
-    MIT, BSD 2-clause, BSD 3-clause revised,
-    Apache 2.0, Lesser GNU General Public
-    License (LGPL), and the
-    GNU General Public License (GPL).
-    Releasing to the public domain (e.g., CC0)
-    meets this criterion if there are no
-    other encumbrances (e.g., patents).
-  control_mappings: # TODO
-  security_insights_value: # TODO
+  - id: OSPS-LE-02
+    maturity_level: 1
+    criterion: |
+      The license for the source code MUST
+      meet the OSI Open Source Definition or
+      the FSF Free Software Definition.
+    rationale: |
+      Ensure that the project's source code is
+      distributed under a recognized and legally
+      enforceable open source software license,
+      providing clarity on how the code
+      can be used and shared by others.
+    details: |
+      Add a LICENSE file to the project's repo
+      with a license that is
+      an approved license by the
+      Open Source Initiative (OSI), or
+      a free license as approved by the
+      Free Software Foundation (FSF).
+      Examples of such licenses include the
+      MIT, BSD 2-clause, BSD 3-clause revised,
+      Apache 2.0, Lesser GNU General Public
+      License (LGPL), and the
+      GNU General Public License (GPL).
+      Releasing to the public domain (e.g., CC0)
+      meets this criterion if there are no
+      other encumbrances (e.g., patents).
+    control_mappings: # TODO
+    security_insights_value: # TODO
 
-- id: OSPS-LE-03
-  maturity_level: 1
-  category: Legal
-  criterion: |
-    The license for the source code MUST be
-    maintained in a standard location within
-    the project's repository.
-  rationale: |
-    Ensure that the project's source code is
-    distributed with the appropriate license
-    terms, making it clear to users and
-    contributors how the code can be used and
-    shared.
-  details: |
-    Include the project's source code license in
-    the project's LICENSE file, COPYING file,
-    or LICENSE/
-    directory to provide visibility and clarity
-    on the licensing terms. The filename MAY
-    have an extension.
-  control_mappings: # TODO
-  security_insights_value: # TODO
+  - id: OSPS-LE-03
+    maturity_level: 1
+    criterion: |
+      The license for the source code MUST be
+      maintained in a standard location within
+      the project's repository.
+    rationale: |
+      Ensure that the project's source code is
+      distributed with the appropriate license
+      terms, making it clear to users and
+      contributors how the code can be used and
+      shared.
+    details: |
+      Include the project's source code license in
+      the project's LICENSE file, COPYING file,
+      or LICENSE/
+      directory to provide visibility and clarity
+      on the licensing terms. The filename MAY
+      have an extension.
+    control_mappings: # TODO
+    security_insights_value: # TODO
 
-- id: OSPS-LE-04
-  maturity_level: 1
-  category: Legal
-  criterion: |
-    The license for the released software assets
-    MUST meet the OSI Open Source Definition or
-    the FSF Free Software Definition.
-  rationale: |
-    Ensure that the project's source code is
-    distributed under a recognized and legally
-    enforceable open source software license,
-    providing clarity on how the code
-    can be used and shared by others.
-  details: |
-    If a different license is included with
-    released software assets, ensure it is
-    an approved license by the
-    Open Source Initiative (OSI), or
-    a free license as approved by the
-    Free Software Foundation (FSF).
-    Examples of such licenses include the
-    MIT, BSD 2-clause, BSD 3-clause revised,
-    Apache 2.0, Lesser GNU General Public
-    License (LGPL), and the
-    GNU General Public License (GPL).
-    Note that the license for the released
-    software assets may be different than the
-    source code.
-  control_mappings: # TODO
-  security_insights_value: # TODO
+  - id: OSPS-LE-04
+    maturity_level: 1
+    criterion: |
+      The license for the released software assets
+      MUST meet the OSI Open Source Definition or
+      the FSF Free Software Definition.
+    rationale: |
+      Ensure that the project's source code is
+      distributed under a recognized and legally
+      enforceable open source software license,
+      providing clarity on how the code
+      can be used and shared by others.
+    details: |
+      If a different license is included with
+      released software assets, ensure it is
+      an approved license by the
+      Open Source Initiative (OSI), or
+      a free license as approved by the
+      Free Software Foundation (FSF).
+      Examples of such licenses include the
+      MIT, BSD 2-clause, BSD 3-clause revised,
+      Apache 2.0, Lesser GNU General Public
+      License (LGPL), and the
+      GNU General Public License (GPL).
+      Note that the license for the released
+      software assets may be different than the
+      source code.
+    control_mappings: # TODO
+    security_insights_value: # TODO
