+++ cmd/template.md
@@ -19,31 +19,42 @@ For more information on the project and to make contributions, visit the [GitHub
 * [Level 2](#level-2): for any code project that has at least 2 maintainers and a small number of consistent users
 * [Level 3](#level-3): for any code project that has a large number of consistent users
 
-### Level 1
 
+### Level 1
+{{ range .Categories }}
 {{- range .Criteria }}
-{{if eq .MaturityLevel 1}}
+{{- if eq .MaturityLevel 1}}
 **[{{ .ID }}]({{ .ID | asLink }})**: {{ .CriterionText | addLinks }}
+{{- end }}
 {{- end }}
 {{- end }}
 
 ### Level 2
-
+{{ range .Categories }}
 {{- range .Criteria }}
-{{if eq .MaturityLevel 2}}
+{{- if eq .MaturityLevel 2}}
 **[{{ .ID }}]({{ .ID | asLink }})**: {{ .CriterionText | addLinks }}
-{{ end }}
+{{- end }}
+{{- end }}
 {{- end }}
 
 ### Level 3
-
+{{ range .Categories }}
 {{- range .Criteria }}
-{{if eq .MaturityLevel 3}}
+{{- if eq .MaturityLevel 3}}
 **[{{ .ID }}]({{ .ID | asLink }})**: {{ .CriterionText | addLinks }}
-{{ end }}
+{{- end }}
+{{- end }}
 {{- end }}
 
-## Criteria Details
+## Categories
+
+{{ range .Categories }}
+
+### {{ .CategoryName }}
+
+{{ .Description }}
+
 
 {{- range .Criteria }}
 
@@ -58,7 +69,7 @@ For more information on the project and to make contributions, visit the [GitHub
 **Category:**
 {{ .Category }}
 
-**Objective:**
+**Rationale:**
 
 {{ .Rationale | addLinks}}
 **Details:**
@@ -80,6 +91,7 @@ _No security insights identified._
 
 ---
 
+{{- end }}
 {{- end }}
